requirements.txt
amqp==5.3.1
asgiref==3.9.1
billiard==4.2.1
celery==5.5.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cron_descriptor==2.0.5
Django==5.2.5
django-celery-beat==2.8.1
django-cors-headers==4.7.0
django-environ==0.12.0
django-filter==25.1
django-timezone-field==7.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
kombu==5.5.4
packaging==25.0
pillow==11.3.0
prompt_toolkit==3.0.52
psycopg==3.2.9
PyJWT==2.10.1
python-crontab==3.3.0
python-dateutil==2.9.0.post0
redis==5.2.1
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
vine==5.1.0
wcwidth==0.2.13
